stage 3: split main.py into 14 sub-routers, preserve handler execution order
